+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.x.x (2017-xx-xx)
+
+## Fixed
+
+- Minor fix and cleanup in `SimpleHttpHandler` channel creation.
+
 ## 1.4.3 (2017-05-20)
 
 ## Fixed